Issue #68: Split Datastore
